add archive sizes to cache
